move bulk create method
