+++ hawc/apps/lit/forms.py
@@ -11,6 +11,7 @@ from ...services.utils import ris
 from ..assessment.models import Assessment
 from ..common.forms import BaseFormHelper, addPopupLink
 from ..common.helper import read_excel
+from ..study.models import Study
 from . import constants, models
 
 logger = logging.getLogger(__name__)
@@ -548,3 +549,35 @@ class ReferenceExcelUploadForm(forms.Form):
                 '"Full text URL", case sensitive.'
             )
         return fn
+
+
+class BulkReferenceStudyExtractForm(forms.Form):
+    class Meta:
+        model = models.Reference
+
+    def bulk_create_studies(cls, study_type: int, reference_ids: list[int], assessment_id: int):
+        """Bulk create studies from a list of reference ids
+        Args:
+            study_type (int): an integer from 0-4 that maps to the study type the objects will have when created
+                0: no study type, 1: bioassay, 2: epi, 3: epi meta, 4: in vitro
+            reference_ids (list[int]): the list of reference ids to convert to studies
+        """
+
+        for reference_id in reference_ids:
+            reference = models.Reference.objects.get(pk=reference_id)
+
+            attrs = {}
+            if study_type == 1:
+                attrs = {"bioassay": True}
+
+            elif study_type == 2:
+                attrs = {"epi": True}
+
+            elif study_type == 3:
+                attrs = {"epi_meta": True}
+
+            elif study_type == 4:
+                attrs = {"in_vitro": True}
+
+            Study.save_new_from_reference(reference, attrs)
+
