+++ hawc/apps/lit/templates/lit/reference_extract_list.html
@@ -25,7 +25,7 @@
             </thead>
             {% for object in object_list %}
                 <tr>
-                    <td><input type="checkbox"></td>
+                    <td><input type="checkbox" name="reference_ids"></td>
                     <td>{{object.authors}}</td>
                     <td>{{object.title}}</td>
                     <td>{{object.year}}</td>
